Separate Esqueleto mtl modules into library
Fixes a TODO in the README.

These Esqueleto modules are meant to be to esqueleto as persistent-mtl
is to persistent, but we don't want them to be provided by this library,
since it's a conflict of responsibility.

Instead, it makes more sense to expose them in a separate library, which
allows other code to use them as necessary.